--- src/containers/components/LoginButton/loginButton.css
@@ -1,7 +1,0 @@
-.buttonStyle {
-    width: 120px;
-    background: #267cc1;
-    border-style: solid;
-    border-width: 1px;
-    color: #fff;
-}
