+++ src/containers/register.js
@@ -83,7 +83,7 @@ class Register extends React.Component {
                     let passwordError = this.props.requisitionError.data.password1
                     if (passwordError.length === 2) {
                         if (passwordError[1] === "Esta senha é inteiramente numérica.") {
-                            inputErrors['password'] += " " + "Não utilize somente números na sua senha."
+                            inputErrors['password'] += " Não utilize somente números na sua senha."
                         } else {
                             inputErrors['password'] += " " + passwordError[1]
                         }
