+++ src/store/reducers/authentication.js
@@ -10,7 +10,7 @@ const Authentication = (state = {}, action) => {
       return {
         ...state,
         isAuthenticated: true,
-        userData: action.payload
+        token: action.payload
       };
     case 'REGISTER_ERROR':
       return {
@@ -27,7 +27,7 @@ const Authentication = (state = {}, action) => {
       return {
         ...state,
         isAuthenticated: true,
-        userData: action.payload
+        token: action.payload
       }
     case 'LOGIN_ERROR':
       return {
@@ -39,7 +39,7 @@ const Authentication = (state = {}, action) => {
       return {
         ...state,
         isAuthenticated: false,
-        userData: undefined
+        token: undefined
       };
     default:
       return state;
